Enable image sync support on TR-FC1-ULAKBIM. (#423)
* feat: Add entitlement for vo.imagine-ai.eu

* Update the list of supported VOs on the site.

This commit does a bit of housekeeping to keep the list of
supported VOs up to date with the catchall repository. In short
the following has been done.

- Remove fusion VO since they have ceased operations.
- Remove ENES VO since they have migrated to another site.
+ Added Protocoast.eu since we're supporting them.

* feat: Enable image sync on TR-FC1-ULAKBIM.

This commit enables image sync support on TR-FC1-ULAKBIM.
Currently only qcow images enabled due to storage constraints.
More formats can be enabled if necessary, or more storage space is acquired.